General (auto) cleanup, make commands work in DMs

+++ src/main/kotlin/net/fabricmc/bot/utils/MessageUtils.kt
@@ -25,7 +25,7 @@ suspend fun Message.respond(content: String): Message = respond { this.content =
  * @param builder Builder lambda for populating the message fields.
  */
 suspend fun Message.respond(builder: MessageCreateBuilder.() -> Unit): Message {
-    val mention = if (this.author != null) {
+    val mention = if (this.author != null && this.getChannelOrNull() !is DmChannel) {
         "${this.author!!.mention} "
     } else {
         ""
@@ -51,7 +51,8 @@ suspend fun Message.respond(builder: MessageCreateBuilder.() -> Unit): Message {
  * the bot commands channel.
  *
  * @param delay How long (in milliseconds) to wait before deleting the response message
- * @param role Minimum role required to bypass the channel requirement, defaulting to trainee moderator
+ * @param role Minimum role required to bypass the channel requirement, defaulting to trainee moderator, or null
+ * to disallow any role bypass
  * @param allowDm Whether to consider DM channels as an appropriate context
  * @param deleteOriginal Whether to delete the original message as well as the response, using the same given delay
  * (true by default)
@@ -59,18 +60,22 @@ suspend fun Message.respond(builder: MessageCreateBuilder.() -> Unit): Message {
  */
 suspend fun Message.requireBotChannel(
         delay: Long = DELETE_DELAY,
-        role: Roles = Roles.TRAINEE_MODERATOR,
+        role: Roles? = Roles.TRAINEE_MODERATOR,
         allowDm: Boolean = true,
         deleteOriginal: Boolean = true): Boolean {
     val botCommands = config.getChannel(Channels.BOT_COMMANDS)
-    val roleObj = config.getRole(role)
+    val roleObj = if (role != null) config.getRole(role) else null
 
-    val topRole = this.getAuthorAsMember()!!.getTopRole()
+    val topRole = if (getGuildOrNull() != null) {
+        this.getAuthorAsMember()!!.getTopRole()
+    } else {
+        null
+    }
 
     @Suppress("UnnecessaryParentheses")  // In this case, it feels more readable
     if (
-            (allowDm && this.channel is DmChannel)
-            || (topRole != null && topRole >= roleObj)
+            (allowDm && this.getChannelOrNull() is DmChannel)
+            || (roleObj != null && topRole != null && topRole >= roleObj)
             || this.channelId == botCommands.id
     ) return true
 
@@ -78,9 +83,39 @@ suspend fun Message.requireBotChannel(
             "Please use ${botCommands.mention} for this command."
     ).deleteWithDelay(delay)
 
-    if (deleteOriginal) {
-        this.deleteWithDelay(delay)
+    if (deleteOriginal && this.getChannelOrNull() !is DmChannel) this.deleteWithDelay(delay)
+
+    return false
+}
+
+/**
+ * Check that this message happened in a guild channel.
+ *
+ * If it didn't, a response message will be created instructing the user that the current command can't be used via a
+ * private message.
+ *
+ * @param role Minimum role required to bypass the channel requirement, defaulting to trainee moderator, or null
+ * to disallow any role bypass
+ * @return true if the message was posted in an appropriate context, false otherwise
+ */
+suspend fun Message.requireGuildChannel(role: Roles? = Roles.TRAINEE_MODERATOR): Boolean {
+    val roleObj = if (role != null) config.getRole(role) else null
+
+    val topRole = if (getGuildOrNull() != null) {
+        this.getAuthorAsMember()!!.getTopRole()
+    } else {
+        null
     }
+
+    @Suppress("UnnecessaryParentheses")  // In this case, it feels more readable
+    if (
+            (roleObj != null && topRole != null && topRole >= roleObj)
+            || this.getChannelOrNull() !is DmChannel
+    ) return true
+
+    this.respond(
+            "This command is not available via private message."
+    )
 
     return false
 }
